+++ src/components/search/BreedFilter/index.tsx
@@ -1,73 +1,9 @@
 import { useState, useEffect } from "react";
-import dogService from "../../services/dogService";
+import dogService from "../../../services/dogService";
+import type { BreedFilterProps } from "./types";
+import BreedOption from "./BreedOption";
+import BreedPill from "./BreedPill";
 
-interface BreedFilterProps {
-  selectedBreeds: string[];
-  onBreedsChange: (breeds: string[]) => void;
-}
-
-function BreedOption({
-  breed,
-  pendingBreeds,
-  toggleBreed,
-}: {
-  breed: string;
-  pendingBreeds: string[];
-  toggleBreed: (breed: string) => void;
-}) {
-  return (
-    <div className="flex items-center p-1 hover:bg-gray-50 rounded">
-      <input
-        id={`breed-${breed}`}
-        type="checkbox"
-        checked={pendingBreeds.includes(breed)}
-        onChange={() => toggleBreed(breed)}
-        className="h-4 w-4 text-primary-600 focus:ring-primary-500 border-gray-300 rounded"
-        onClick={(e) => e.stopPropagation()}
-      />
-      <label
-        htmlFor={`breed-${breed}`}
-        className="ml-2 block text-sm text-gray-700 truncate"
-      >
-        {breed}
-      </label>
-    </div>
-  );
-}
-
-function PendingBreedPill({
-  breed,
-  toggleBreed,
-}: {
-  breed: string;
-  toggleBreed: (breed: string) => void;
-}) {
-  return (
-    <span
-      key={breed}
-      className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-primary-100 text-primary-800"
-    >
-      {breed}
-      <button
-        type="button"
-        className="ml-1.5 inline-flex text-primary-600 hover:text-primary-900"
-        onClick={(e) => {
-          e.stopPropagation();
-          toggleBreed(breed);
-        }}
-      >
-        <span className="sr-only">Remove</span>
-        <svg className="h-3 w-3" fill="currentColor" viewBox="0 0 20 20">
-          <path
-            fillRule="evenodd"
-            d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z"
-            clipRule="evenodd"
-          />
-        </svg>
-      </button>
-    </span>
-  );
-}
 
 function BreedFilter({ selectedBreeds, onBreedsChange }: BreedFilterProps) {
   const [breeds, setBreeds] = useState<string[]>([]);
@@ -256,7 +192,7 @@ function BreedFilter({ selectedBreeds, onBreedsChange }: BreedFilterProps) {
                   </div>
                   <div className="flex flex-wrap gap-2 mb-3">
                     {pendingBreeds.map((breed) => (
-                      <PendingBreedPill
+                      <BreedPill
                         key={breed}
                         breed={breed}
                         toggleBreed={toggleBreed}
