+++ src/components/search/BreedFilter/types.ts
@@ -1,0 +1,4 @@
+export interface BreedFilterProps {
+  selectedBreeds: string[];
+  onBreedsChange: (breeds: string[]) => void;
+}
